moved comparison operation tests into a separte file
